k8s-mcp-server/k8s-mcp-server/pyproject.toml
[project]
name = "k8s-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.24.0",
    "kubernetes>=32.0.1",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.4.3",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
]
